[FIX] Prevent null object reference being deleted

+++ app/src/main/java/com/example/androidherexamen/main/MainViewModel.kt
@@ -24,6 +24,7 @@ class MainViewModel(userId: Int, val database: PostDatabaseDAO, application: App
     }
 
     private suspend fun delete(post: Post){
+        if(post != null)
         database.delete(post)
     }
